requirements.txt
Django==4.2.24
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
Pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-ratelimit==4.1.0
dj-database-url==2.1.0
pytest==7.4.3
pytest-django==4.7.0
gunicorn==21.2.0
# ECPay Payment Integration
requests==2.31.0
pycryptodome==3.19.0